HomeWork7.2
